Finish todo by user

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth.actions';
 import prisma from '@/lib/prisma';
 import { NextResponse, NextRequest } from 'next/server';
 import * as yup from 'yup';
@@ -25,10 +26,13 @@ const postSchema = yup.object({
 })
 
 export async function POST(request: Request) {
+  const user = await getUserSessionServer();
+  if(!user) NextResponse.json('Not authorized', {status:401});
+
   try {
     const {complete,description} = await postSchema.validate(await request.json());
     const todo = await prisma.todo.create({
-      data:{complete,description}
+      data:{complete,description,userId: user?.id}
     });
   
     return NextResponse.json({todo});
@@ -39,8 +43,11 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(_: Request) {
+  const user = await getUserSessionServer();
+  if(!user) NextResponse.json('Not authorized', {status:401});
+
   try {
-    await prisma.todo.deleteMany({where:{complete:true}});
+    await prisma.todo.deleteMany({where:{complete:true, userId: user?.id}});
     return NextResponse.json({success:true});
     
   } catch (error) {
